requirements.txt
Django==4.2.0
stellar-sdk==10.0.0
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
dj-database-url==2.1.0
psycopg2-binary==2.9.9
